Moved files in TableComponents folder
