+++ src/components/TableComponents/index.ts
@@ -2,4 +2,4 @@ export * from './ActivityCells'
 export * from './CollaterallCell'
 export * from './CommonCells'
 export * from './HeaderCell'
-export * from './cells'
+export * from './helpers'
